256: change over to github actions (#257)
* Rework CI

Moved from docker images and circle CI to github actions. Lost the ability to run it locally as it runs in github actions for now, but otherwise it seems better. During the rework, broke the problem into two parts: static and dynamic.

The static analysis does everything from code style (black) to analysis (pylint). Many tools are used for each of the checks including some home grown tools for dawgie compliance.

The dynamic bit is pytest with coverage to be added in the future. The current coverage checks never worked all that well, but will to try it again with new system. The dynamic can also expand to include larger more complex integrated tests if need be.