update TLS internal usage
To use TLS, myself now has to be define. If dawgie.context.ssl_pem_file is defined and exists, it will be used as the PEM for all listening sockets overriding myself. This allows the user to have ad-hoc certs for themself and use it to authenticate but have a chained cert for the listener so it can be looked up and validated if desired by the client.