repair how dawgie uses git
DAWGIE wants to use 3 branches. The first branch is the stable branch to pull changesets from. It does this via a simple pull and will jump to the head of the stable branch.

The second branch is an ops testing branch. When a submit is requested, the stable branch is updated and a test branch created at the head. It then does dawgie.tools.compliant.

Once the test branch has shown that stable HEAD is dawgie.tools.compliant, it deletes the current ops branch and then creates it again (the third branch) at the new stable HEAD. It is best to have this branch created before startng DAWGIE.

Moved some useful parameters from dawgie.tools.submit to dawgie.context for consistency and ease of use. In moving them, gave them environment names and CLI arguments.