277: archive is an array of strings (#278)
* archive is an array of strings and not handled properly

Use last value given. The list of strings is an artifact from the DynamicContent and that it does not know how to handle multiple args in the same request. Therefore it defaults to making a list of all args and putting them in order and letting the specific instance of DynamicContent worry about down converting from list to single item when necessary.

Changed how and when archiving is triggered. Just need to pretend that there is new data by setting dawgie.pl.farm.ARCHIVE and then doing the reset. Calling fsm.trigger_archiving() directly is the wrong way.