adjust messaging and hold off
Reviewing the log indicates that can trap for the unique error when doing updates and change to a simple warning. There was one case where it looked like that same tasks collided multiple times. Put a uniform distribution draw of a sleep in place to try and prevent an infinite loop of two perfectly synchronzied tasks.