update dawgie.db.post.metrics()
Have it be lazy on loading the actual data. Instead of loading everything when it may not be used, load just the bits when asked for it. Quick review of navel gazing suggests that it loads half the data. The next step would be to make it smarter and load only what it has not seen before. If done at reboot, this will still be everything. At reload, it should be a tiny subset.