update dawgie.db.post.metrics() (#293)
* update dawgie.db.post.metrics()

Have it be lazy on loading the actual data. Instead of loading everything when it may not be used, load just the bits when asked for it. Quick review of navel gazing suggests that it loads half the data. The next step would be to make it smarter and load only what it has not seen before. If done at reboot, this will still be everything. At reload, it should be a tiny subset.

* add unit test for LazyMetricValue

Show that it does the correct job at unpickling the dawgie.Value when the content is requested.

* rewrote resources to leverage lazy loading

Have dawgie keep the results of navel gazing in a diary,pkl. Check the file contents after loading to make sure they look as expected. Should improve navel gazing times. Used live data to test it.