change error handling
The Deferred was running the fail exception down the chain despite being handled as programmed in dawgie.fe.submit and documented in twisted promising a pipeline failure due to unhandled exception when any exception occurred.

Put a flag into dawgie.fe.submit.Process() that is marked when a failure occurs. Once set, callbacks return None without doing anything. This should result in the completion of the Deferred stack without an unhandled exception at the end.