Issue 295: unique entries in prime table (#296)
* update self diagnostics (unit/regression tests)

Once the table was unique, the unit tests failed. Cleaned up the testdata to make them all unique. Also found some bad testing for remove() and update(). Fixed those tests as well.

Added a new test for dawgie.Regressions.

* make entries unique

When the prime table is created, rows now have to unique based on run id, target, task, alg, sv, and value. Because of the uniqueness and that regressions are always run id 0, modified the update to do an 'upsert' for run id 0 only. All other duplicates will generate a uniqueness failure.