two prong fix (#300)
In a previous request, navel gazing was made threaded since our pipeline does not depend on the insights it grants (select between cluster or aws). In the case that it would, it uses the hint provided by the module to be run when there are no insights. The multithreaded change set up the conditions that requires this fix.

1. The read assumes the pickle file is readable and prevents navel gazing from completing if the pickle.load() throws an exception. Trap the known exception, UnpicklingError, during read to allow navel gaze to complete with some insights.

2. Protect the writing of the insights to the diary by writing to a temporary file first, then, when successful, copy it to the directory to reduce the problem of killing navel gazing during the write.

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>