basics are in place
Running Coda/exercise/run.sh seems to be working well enough. It builds the images, starts postgres followed by the pipeline. It does not do the actual exercise yet because need to add some workers. Should also add a line in run.sh that waits for the user to hit return before shutting down and cleaning up. Maybe do it if an env is defined.