Issue 247: shelve does not work with exerciese (#301)
* First, restructured a small bit of the code

Moved Python/Coda to Coda. While it used to be the Coda for just dawgie and AWS it has broadened to included the run.sh for emulating the github actions. Added Coda/exercise.

* Second, moved the exercise code from the past to docker compose

Coda/exercise/run.sh is the entry to build then do the exercise routines. It uses a multi stage build, defaulting to stable, for the base docker image. The server and workers build from the base basically adding just the entry point for each task.

Docker compose will then load the appropriate postgresql image and run it. Once postgresql is healthy, the pipeline will start. The delay happens even if the shelve DB is being used instead of postgresql. Once the pipeline is up and healthy, the workers will be created and then recreated indefinitely.

Once the full system is up and running, Coda/exercise/run.sh will request that the feedback tree be exercised by commanding the pipeline to run feedback.command. This tree runs indefinitely.

The pipeline is accessible while Coda/exercise/run.sh is running. Coda/exercise/run.sh cleans up after itself through the normal exit or a ctrl-c.

*  Third and last, shelve was update to correct minor issues

All tables need to be updated from anywhere for proper pipeline behavior. It required changing workers calling dawgie.db.shelve.update() to send requests back to the mother ship.

The names being used in bot.new_values() were wrong. shelve was passing back a string of the ids tuple. Changed to be the string of the names delimited by '.'.