update logging to help find problem (#305)
It seems that targets for esp that contain the plus sign (+) are failing in view. The original thought was that the sql statement was weirdly interpreting the plus sign. A quick test revealed it was not the case. Improved the statements to show the actual target and recorded the actual exception in the log files.