WIP: initial transition monitoring
Updated to latest transitions library as well as postgres for this work. Had to do both as newest transitions was causing problems with psycopg. Still have pylint problems with psycopg but worked a meh solution.

Now have the ability to monitor transitions but have to manually add them in the call backs.