Issue 302: FSM problems (#303)
* cleaned up unit testing and added baseline tests for FSM

Improved some of the mock items cleaning up the unit testing just a tad. Added a baseline change of state set of tests that verify transitions and such from state to state.

* added transition guards

The biggest changes came in adding transition guards that do not allow for state transitions to take place while in the FSM is entering/exiting states. It is the one change that directly addresses the base problem as it seemed that transitions were requested before fully in the given state. While this would not happen with a traditional synchronous FSM, the FSM used is synchronous but the enter/exit functions are twisted and, as such, asynchronous. The guards will block badly ordered triggers().