Issue 310: improving boot times (#311)
* Repaired the FSM

Removed the navel gazing state because it was taking so long and tried to make it asynchronous. This causes problems elsewhere when states are changed by the user with a reset. So, back into the state machine.

* Reduce processing time

While processing time for navel gazing is fixed by the number of metric data values (50 ms per state vector or 14 values), it was possible to clean up the navel gazing to properly load the diary it has been keeping.

* Reduce loaded metrics

The next tall tent pole in the processing is retrieving millions of metrics when only a few hundred are new. By find the latest run ID in the diary, can have the DB retrieve just the newer run IDs and run ID 0 since they are overwritten.